Fix kernel boot issues and complete interrupt system - VGA loop fix - Clean entry point - Fix panic noreturn warning - Full PIC/IDT/keyboard interrupt support

+++ kernel/vga.c
@@ -26,7 +26,7 @@ void vga_init(void) {
 
 /* Clear the entire screen */
 void vga_clear(void) {
-    for (uint16_t i = 0; i < VGA_WIDTH * VGA_HEIGHT; i++) {
+    for (uint32_t i = 0; i < (uint32_t)(VGA_WIDTH * VGA_HEIGHT); i++) {
         VGA_BUFFER[i] = vga_entry(' ', VGA_COLOR_LIGHT_GRAY);
     }
     cursor_x = 0;
